Refactor element operation details to use string formatting for improved readability and consistency across various operations. Update search criteria handling in element search operations to simplify and enhance clarity. Introduce a new method for creating search criteria strings. Add a README for UIAutomationMCP tests outlining safety policies, test categories, and execution guidelines.

+++ UIAutomationMCP.Worker/Operations/Window/WaitForInputIdleOperation.cs
@@ -75,7 +75,7 @@ namespace UIAutomationMCP.Worker.Operations.Window
                     ActionName = "WaitForInputIdle",
                     Completed = success,
                     TimeoutMilliseconds = timeoutMilliseconds,
-                    ElapsedMilliseconds = elapsed.TotalMilliseconds,
+                    ElapsedMilliseconds = (int)elapsed.TotalMilliseconds,
                     TimedOut = !success,
                     WindowInteractionState = windowPattern.Current.WindowInteractionState.ToString(),
                     Message = success
